T 07.01.2025 04:21

+++ NN/Main.cpp
@@ -2,7 +2,7 @@
 #include "NeuralNetwork.h"
 
 int main() {
-	srand(time(0));
+	srand((unsigned int)time(0));
 
 	addit f;
